kafka_consumer/requirements.txt
kafka-python
psycopg2-binary
websocket-client
python-dotenv
confluent_kafka